feat: update resources page
