+++ web-www/components/templates/resources/resourcesSection/resourcesSection.styles.ts
