--- web-www/graphql/sanity/fragments/resources/resourcesLedgerSection.graphql
@@ -1,16 +1,0 @@
-fragment resourcesLedgerSectionFields on ResourcesPage {
-  ledgerSection {
-    header
-    cards {
-      titleRaw
-      descriptionRaw
-      lastUpdated
-      button {
-        ...buttonFields
-      }
-      image {
-        ...customImageFields
-      }
-    }
-  }
-}
