--- web-www/graphql/sanity/fragments/resources/resourcesRegistrySection.graphql
@@ -1,26 +1,0 @@
-fragment resourcesRegistrySectionFields on ResourcesPage {
-  registrySection {
-    header
-    documentTableTitle
-    documents {
-      name
-      type
-      date
-      href
-    }
-    subsections {
-      title
-      cards {
-        image {
-          ...customImageFields
-        }
-        titleRaw
-        lastUpdated
-        descriptionRaw
-        button {
-          ...buttonFields
-        }
-      }
-    }
-  }
-}
