+++ web-www/graphql/sanity/fragments/shared/resourceFieldsFragment.graphql
@@ -1,0 +1,14 @@
+fragment resourceFields on ResourcesSection {
+  header
+  cards {
+    titleRaw
+    descriptionRaw
+    lastUpdated
+    button {
+      ...buttonFields
+    }
+    image {
+      ...customImageFields
+    }
+  }
+}
