+++ web-www/graphql/sanity/pages/resourcesPage.graphql
@@ -1,7 +1,6 @@
 query resourcesPage {
   allResourcesPage {
     ...resourcesTopSectionFields
-    ...resourcesRegistrySectionFields
-    ...resourcesLedgerSectionFields
+    ...resourcesSectionsFields
   }
 }
