+++ web-www/pages/_app.tsx
@@ -8,6 +8,7 @@ import theme from 'web-components/lib/theme/muiTheme';
 
 import '../styles/font.css';
 import '../styles/web-components.css';
+import '../styles/background.css';
 import Layout from '@/components/templates/Layout/Layout';
 
 const { augmentDocumentWithEmotionCache, withAppEmotionCache } =
